Move effect related packet handlers to Effects namespace/folder
