+++ EOLib/EOLib.csproj
@@ -253,16 +253,16 @@
     <Compile Include="PacketHandlers\Commands\FindCommandPlayerSameMapHandler.cs" />
     <Compile Include="PacketHandlers\Chat\GroupChatHandler.cs" />
     <Compile Include="PacketHandlers\DoorOpenHandler.cs" />
-    <Compile Include="PacketHandlers\EffectPotionHandler.cs" />
+    <Compile Include="PacketHandlers\Effects\EffectPotionHandler.cs" />
     <Compile Include="PacketHandlers\EndPlayerWarpHandler.cs" />
     <Compile Include="PacketHandlers\ItemPickupHandler.cs" />
     <Compile Include="PacketHandlers\MainPlayerWalkHandler.cs" />
-    <Compile Include="PacketHandlers\MapHpDrainHandler.cs" />
+    <Compile Include="PacketHandlers\Effects\MapHpDrainHandler.cs" />
     <Compile Include="PacketHandlers\NPCTakeDamageHandler.cs" />
     <Compile Include="PacketHandlers\PlayerTargetOtherSpellHandler.cs" />
     <Compile Include="PacketHandlers\PlayerSelfTargetSpellHandler.cs" />
-    <Compile Include="PacketHandlers\PlayerSpikeDamageHandler.cs" />
-    <Compile Include="PacketHandlers\MapDebuffHandler.cs" />
+    <Compile Include="PacketHandlers\Effects\PlayerSpikeDamageHandler.cs" />
+    <Compile Include="PacketHandlers\Effects\MapDebuffHandler.cs" />
     <Compile Include="PacketHandlers\NPCActionHandler.cs" />
     <Compile Include="PacketHandlers\NPCEnterMapHandler.cs" />
     <Compile Include="PacketHandlers\NPCLeaveMapHandler.cs" />
@@ -317,7 +317,7 @@
     <Compile Include="PacketHandlers\RecoverStatListHandler.cs" />
     <Compile Include="PacketHandlers\RefreshMapStateHandler.cs" />
     <Compile Include="PacketHandlers\StatTrainingHandler.cs" />
-    <Compile Include="PacketHandlers\TimedSpikeEffectHandler.cs" />
+    <Compile Include="PacketHandlers\Effects\TimedSpikeEffectHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Win32.cs" />
     <Compile Include="HDSerialNumberServiceLinux.cs" />
