+++ EOLib/PacketHandlers/Effects/MapDebuffHandler.cs
@@ -7,7 +7,7 @@ using EOLib.Net.Handlers;
 using System;
 using System.Collections.Generic;
 
-namespace EOLib.PacketHandlers
+namespace EOLib.PacketHandlers.Effects
 {
     [AutoMappedType]
     public class MapDebuffHandler : InGameOnlyPacketHandler
